+++ stack/dokku.py
@@ -1,7 +1,7 @@
 import troposphere.cloudformation as cloudformation
 import troposphere.ec2 as ec2
 import troposphere.iam as iam
-from troposphere import Base64, FindInMap, Join, Output, Parameter, Ref, Tags
+from troposphere import Base64, FindInMap, Join, Output, Ref, Tags
 from troposphere.policies import CreationPolicy, ResourceSignal
 
 from .assets import assets_management_policy
@@ -10,6 +10,7 @@ from .domain import domain_name
 from .environment import environment_variables
 from .logs import logging_policy
 from .template import template
+from .utils import ParameterWithDefaults as Parameter
 from .vpc import container_a_subnet, vpc
 
 key_name = template.add_parameter(