+++ stack/environment.py
@@ -2,7 +2,14 @@ import os
 
 from troposphere import AWS_REGION, GetAtt, If, Join, Ref
 
-from .assets import assets_bucket, distribution, private_assets_bucket
+from .assets import (
+    assets_bucket,
+    assets_cloudfront_domain,
+    assets_custom_domain_condition,
+    assets_use_cloudfront_condition,
+    distribution,
+    private_assets_bucket
+)
 from .cache import (
     cache_cluster,
     cache_condition,
@@ -10,7 +17,14 @@ from .cache import (
     using_redis_condition
 )
 from .common import secret_key
-from .database import db_condition, db_instance, db_name, db_password, db_user
+from .database import (
+    db_condition,
+    db_engine,
+    db_instance,
+    db_name,
+    db_password,
+    db_user
+)
 from .domain import domain_name, domain_name_alternates
 
 if os.environ.get('USE_GOVCLOUD') != 'on':
@@ -30,12 +44,15 @@ environment_variables = [
     ("DATABASE_URL", If(
         db_condition,
         Join("", [
-            "postgres://",
+            Ref(db_engine),
+            "://",
             Ref(db_user),
             ":",
             Ref(db_password),
             "@",
             GetAtt(db_instance, 'Endpoint.Address'),
+            ":",
+            GetAtt(db_instance, 'Endpoint.Port'),
             "/",
             Ref(db_name),
         ]),
@@ -65,7 +82,16 @@ environment_variables = [
 if distribution:
     # not supported by GovCloud, so add it only if it was created
     environment_variables.append(
-        ("CDN_DOMAIN_NAME", GetAtt(distribution, "DomainName"))
+        ("CDN_DOMAIN_NAME", If(
+            assets_use_cloudfront_condition,
+            If(
+                # use the custom domain passed into the stack, otherwise fallback to the default domain
+                assets_custom_domain_condition,
+                Ref(assets_cloudfront_domain),
+                GetAtt(distribution, "DomainName"),
+            ),
+            "",
+        )),
     )
 
 if es_domain: