+++ stack/template.py
@@ -80,6 +80,7 @@ template.set_group_order([
     'Global',
     'Application Server',
     'Load Balancer',
+    'Static Media',
     'Database',
     'Cache',
     'Elasticsearch',