+++ CHANGELOG.rst
@@ -1,6 +1,25 @@
 Change Log
 ==========
 
+`1.3.0`_ (2018-09-13)
+---------------------
+
+Features:
+
+* Allow overriding parameter defaults at template creation time without having to change the
+  Python code.  See `the README
+  <https://github.com/caktus/aws-web-stacks/blob/master/README.rst#dokku>`_.
+* Add a parameter to control whether certificates are validated by DNS or email, and default
+  to DNS since GDPR has made email validation less likely to work.
+* The database type of the RDS instance can now be configured (previously, only Postgres could
+  be used). Note that, for backwards-compatibility reasons, the resources in the CloudFormation
+  stack is still named ``PostgreSQL`` (this avoids unnecessarily recreating the RDS instance
+  on pre-existing stacks). See: PR #32
+* The RDS instance now supports all allowable special characters in the password field. See: PR #31
+* The CloudFront distribution linked to the S3 assets bucket can now be disabled / enabled at the
+  time a stack is created or updated; the CloudFront distribution now supports a custom domain name
+  and SSL certificate. See: PR #30
+
 
 `1.2.0`_ (2017-09-27)
 ---------------------
@@ -115,6 +134,7 @@ Backwards-incompatible changes:
 * Initial public release
 
 
+.. _1.3.0: https://aws-web-stacks.s3.amazonaws.com/index.html?prefix=1.3.0/
 .. _1.2.0: https://aws-web-stacks.s3.amazonaws.com/index.html?prefix=1.2.0/
 .. _1.1.2: https://aws-web-stacks.s3.amazonaws.com/index.html?prefix=1.1.2/
 .. _1.1.1: https://aws-web-stacks.s3.amazonaws.com/index.html?prefix=1.1.1/